remove estimated_size field

+++ mining/src/mempool/config.rs
@@ -1,7 +1,7 @@
 use kaspa_consensus_core::constants::TX_VERSION;
 
 pub(crate) const DEFAULT_MAXIMUM_TRANSACTION_COUNT: usize = 1_000_000;
-pub(crate) const DEFAULT_MEMPOOL_SIZE_LIMIT: u64 = 500_000_000;
+pub(crate) const DEFAULT_MEMPOOL_SIZE_LIMIT: usize = 500_000_000;
 pub(crate) const DEFAULT_MAXIMUM_BUILD_BLOCK_TEMPLATE_ATTEMPTS: u64 = 5;
 
 pub(crate) const DEFAULT_TRANSACTION_EXPIRE_INTERVAL_SECONDS: u64 = 24 * 60 * 60;
@@ -28,7 +28,7 @@ pub(crate) const DEFAULT_MAXIMUM_STANDARD_TRANSACTION_VERSION: u16 = TX_VERSION;
 #[derive(Clone, Debug)]
 pub struct Config {
     pub maximum_transaction_count: usize,
-    pub mempool_size_limit: u64,
+    pub mempool_size_limit: usize,
     pub maximum_build_block_template_attempts: u64,
     pub transaction_expire_interval_daa_score: u64,
     pub transaction_expire_scan_interval_daa_score: u64,
@@ -52,7 +52,7 @@ impl Config {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         maximum_transaction_count: usize,
-        mempool_size_limit: u64,
+        mempool_size_limit: usize,
         maximum_build_block_template_attempts: u64,
         transaction_expire_interval_daa_score: u64,
         transaction_expire_scan_interval_daa_score: u64,
